refactor(endpoint): move combinators to ext/
